created user.kt
moved content of package user into user.kt
