+++ persistence/src/main/kotlin/com/github/jactor/persistence/ApiExtensions.kt
@@ -1,7 +1,6 @@
 package com.github.jactor.persistence
 
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.user.UserModel
 import com.github.jactor.shared.api.PersistentData
 import com.github.jactor.shared.api.PersonDto
 import com.github.jactor.shared.api.UserDto
