+++ persistence/src/test/kotlin/com/github/jactor/persistence/BlogRepositoryTest.kt
@@ -5,7 +5,6 @@ import java.util.UUID
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.user.UserModel
 import assertk.assertAll
 import assertk.assertThat
 import assertk.assertions.hasSize
