+++ persistence/src/test/kotlin/com/github/jactor/persistence/GuestBookRepositoryTest.kt
@@ -5,9 +5,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.fail
 import org.springframework.beans.factory.annotation.Autowired
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.user.UserModel
-import com.github.jactor.persistence.user.UserBuilder
-import com.github.jactor.persistence.user.UserRepository
 import assertk.assertAll
 import assertk.assertThat
 import assertk.assertions.isEqualTo
