+++ persistence/src/test/kotlin/com/github/jactor/persistence/GuestBookServiceTest.kt
@@ -4,7 +4,6 @@ import java.util.Optional
 import java.util.UUID
 import org.junit.jupiter.api.Test
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.user.UserModel
 import assertk.assertAll
 import assertk.assertThat
 import assertk.assertions.hasSize
