+++ persistence/src/test/kotlin/com/github/jactor/persistence/PersonRepositoryTest.kt
@@ -4,8 +4,6 @@ import java.util.UUID
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.user.UserBuilder
-import com.github.jactor.persistence.user.UserModel
 import assertk.assertAll
 import assertk.assertThat
 import assertk.assertions.contains
