+++ persistence/src/test/kotlin/com/github/jactor/persistence/UserModelTest.kt
@@ -1,10 +1,9 @@
-package com.github.jactor.persistence.user
+package com.github.jactor.persistence
 
 import java.time.LocalDateTime
 import java.util.UUID
 import org.junit.jupiter.api.Test
 import com.github.jactor.persistence.common.PersistentModel
-import com.github.jactor.persistence.PersonModel
 import assertk.assertAll
 import assertk.assertThat
 import assertk.assertions.isEqualTo
