+++ persistence/src/test/kotlin/com/github/jactor/persistence/aop/ModifierAspectTest.kt
@@ -16,8 +16,8 @@ import com.github.jactor.persistence.GuestBookEntryModel
 import com.github.jactor.persistence.GuestBookModel
 import com.github.jactor.persistence.PersonBuilder
 import com.github.jactor.persistence.PersonModel
-import com.github.jactor.persistence.user.UserBuilder
-import com.github.jactor.persistence.user.UserModel
+import com.github.jactor.persistence.UserBuilder
+import com.github.jactor.persistence.UserModel
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isStrictlyBetween
